Added LoL packager
Tested with EU West version, worked fine.
Only problem found: the downloader can crash, running out of file
descriptors. I'm not sure something can be done about it (if there's a
soft ulimit lower than the hard limit maybe it can be raised, but that's
about it).
If you restart the packager script with "overwrite", the download can go
on and the packaging should complete where it left.